Added support for webhook handlers for stripe
Stripe Webhook workflow

- updated AdminAuthorizedKey to map to SITE_ADMIN_AUTHORIZED KEY. this key
allows comms between authenticated servers and acts as a protective barrier
between cross app commmunications.

- Updated webhook handler to properly support card payment and bank account
ACH payment systems. This webhook allows us to listen into stripe events in real
time and update payments made out on the fly.

- Added session id and other variables to 0007 checkout session function.

- Added handlers to update payments based on webhook handlers

Tech debt cleanup

- added required env variables in the env.sample file under dev

- removed un-used rtk query for creating a payment intent since we
no longer want to use our front end to manage that and rather allow
the cloud functions to properly handle the workflow.

- removed unused mutation functions and removed the ability to create
payment records from the ui in PropertyOwnerInfoCard.js file.

- removed the ability to confirmPayment via hook since we do not need that.

- disabled the use of 0009_fetch_stripe_payment_confirmation.js file since we
do not use it anymore since it was legacy way of fetching and ensuring that
the data is correct and up to date.

- Added comments and refactored function isFlagEnabled to view
if a feature flag is enabled easily.

- Updated code to easily be able to view feature flags and if they are avaiable or not.

- Updated MyRental.jsx and Property.jsx since we had a lot of common components between them.
Decided that we should seperate out into propertyDetails and propertyHeader components. When we
get a chance, we should also lace these with jest unit tests. this should give our work a pretty
solid coverage as well.

- Consolidated code for Quick Acctions menu as well.

- Noticed that we have bunch of components just laying around that could might have been better
with some cleanup. We should make it our top priority since we are almost done with the webhook and
we just need to attach the landing page and push the new version to prod.

- added cleanup for associate tenants popup component

- Adding webhook, we wanted to test the api call locally. In order to
do this, we need to install stripe locally so we can test the webhook
properly.

- fixed issue where tenant could create a property by directly navigating towards the
properties route. this fix allows us to prevent that. Based on the roleType the routes
are built now.

- updated dev steps to run netlify functions on debug mode. This
ensures that we can test the webhook handlers locally as well.

- updated stripe workflow to take correct amount into consideration. Previously
we were attempting to convert the NaN values to number all over the place, and we
consolidated the work at one spot.

Closes #94