Tech debt refactor
- Updated MyRental.jsx and Property.jsx since we had a lot of common components between them.
Decided that we should seperate out into propertyDetails and propertyHeader components. When we
get a chance, we should also lace these with jest unit tests. this should give our work a pretty
solid coverage as well.

- Consolidated code for Quick Acctions menu as well.

- Noticed that we have bunch of components just laying around that could might have been better
with some cleanup. We should make it our top priority since we are almost done with the webhook and
we just need to attach the landing page and push the new version to prod.