Added help and support back on the navigation menu
Previously, the help and support was not displaying correctly. Because of this, we were not
able to display help and support on release v1.0.3 . Either we need to patchfix this into v1.0.3 or
we should create new v1.0.4

Also extended this support to individual property pages as well. Previously, due to some weird route
issue we were not able to see the correct details for selected property. Should be fixed as a part
of this ticket.

Closes #129