Updated Splash page
- Updated splash page with ability to click and navigate to respective apps. Rental App
however needs login priv and hence if the client logs in via the tile aka the Rental App Card in the landing page, they are logged in as a property owner automatically. To access their rental account, the client can login via the button on the top.

We also decided that we should keep the button to allow property managers to login in from the top also present since we do need clearer distinction between them two.

Furthermore, this is a design issue that probably needs addressing. The problem here is we don't really care if you are a tenant or owner, we should display stuffs relative to what you have. And we do that to some degree. Removing roles fully without understanding the scope might be difficult to address. Lets keep updating this and fix it as we go.