Updated Templates and Settings workflows
- Added some minor improvement and bug fixes for Rent App.

- The forms fields were not looking so great in small
form factor, hence the update to fix that.

- Removes default dayjs values

Removed custom font to support system font

- After a lot of talks, we realized that using system font
is the best option for our use case. Since we do not want to
spoil user experience and we want a fast rendering of user data
we choose to use the system ui font. Adds variation to macOs

Restructured Invoice Templates in Invoice App

- Updated the look and feel of invoice templates within the invoice app.

Updated templates to better suite new ux design

Although we had a really good ux design, it was not supported
quite right with our users. So in order to properly setup the
templates, we came up with a new design that i'm hoping that it
will support what we want.

Update Template processor

Closes #142