Added Esign capabilities in web
Updated "Manage payments" selection tab under settings

- added ability to choose between stripe and esign details under settings.
Added tabPanel component to support this ask. Updated esign details to match
pretty much StripeConnect exactly. Couple of unknowns here as well.

- Unify Help and Support page

Added a single help and support page for both stripe connect and esign

- Added Recent Documents for Docusign

Update some minor missing fields

- Created rtk query for all external integrations api to use later
down the road.
- Updated Esign Connect with some minor css cleanup.

- Updated TabPanel.jsx to support proper icon display when in small screen
vs when in large screen.

- Updated help and support for better verbiage

- Encountered snag with docusign, moved to SignRequest

What we didn't forsee is that the cost of using docusign is very steep. This unfortunately does not
come even close to the expectation of what we need to deliver. After a quick find, we have landed with
SignRequest Api for our docusign process since they are cheaper alternative.

After this realization, we moved on to making the Esign Connect better. We renamed it to better match
the use case of signing documents and cleared up and existing link with Docusign. In the future, if client
requests increase, we can pivot to docusign or any other external integration. At the end of the day robust
platforms are built like this.

Updated tour steps for external integrations

Disabled esign until backend