Removed all instance of user roles
- After the new change in the router configuration, we were finally able to
get the app to the level where we didn't need roles and that is where we have
arrived. The idea is to allow users to be the manager of their own system, however
we still only support one tenant view aka a tenant cannot be the same across two properties.