Added banner for Demo User Account
update tests

Added roles into webApp

- Added roles in web. Allocated Roles are attributes that can be allocated to the user.
During auth, a basic role of Consumer Demo is allocated. Administrators are the only authorized
person to update the db for other various roles.

- Updated AuthenticationProvider component.
When a user is visiting a page that requires auth, we fetch userData from the db. If the user identity
has been scambled with in the localStorage we catch the error and we redirect the user to home screen.

Closes #209

Added changes to support roles

- Accomplished the task of role for user level by providing no roles for clients who log in. The administrator
is responsible to provide the role. Once a role is provided, the banner is removed from the Rental App A/o.

Closes #209

Added react-secure-storage

- Added react-secure-storage to store user data a bit securely.

- Added jests tests for components.

setup correct timers for jest and dayjs utc

update more jest tests across the webApp

Added more fixes to streamline jest tests

- The problem is our webApp displays text such as "xx hours ago". Unfortunately,
passing date results in dayjs() thinking that it get incremental. Unfortunately,
jest tests cannot know if its 8 hours ago or 7 hours ago and other tests would have
been fine, but snapshot tests are different. Hence, we are having the work the
tests out in a way where it does not impact the suite in any run mode.