Added roles into webApp
- Added roles in web. Allocated Roles are attributes that can be allocated to the user.
During auth, a basic role of Consumer Demo is allocated. Administrators are the only authorized
person to update the db for other various roles.

- Updated AuthenticationProvider component.
When a user is visiting a page that requires auth, we fetch userData from the db. If the user identity
has been scambled with in the localStorage we catch the error and we redirect the user to home screen.

Closes #209