Update authentication workflow
- onAuthStateChanged function now can handle role from auth package
instead of using localstorage. a user without any role is automatically
a regular user and is considered a demo user. we also check if the user
has any invites pending and if such is the case, then we create the user
under that name for auto login purposes.

- created associatedTenant mutation fn where we combine multiple mutations
to enhance code quality and performance.

- tech debt and code clean up.

Closes #209